Cleaned up some of the code when running tests on test functions.
